// frontend/lib/client/types.gen.ts
export type WorkflowActionResponse = {
  /**
   * Id
   */
  id: string
  /**
   * Position
   */
  position: number
  /**
   * Action Type
   */
  action_type: string
  /**
   * Description
   */
  description: string
  /**
   * Payload
   */
  payload: {
    [key: string]: unknown
  }
  /**
   * Status
   */
  status: string
  /**
   * Result Url
   */
  result_url: string | null
  /**
   * Error
   */
  error: string | null
}